pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "newton"
dynamic = ["version"]
license = "Apache-2.0"
license-files = ["LICENSE.md", "newton/licenses/**/*.txt"]
authors = [{ name = "Newton Developers", email = "warp-python@nvidia.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Environment :: GPU :: NVIDIA CUDA",
    "Environment :: GPU :: NVIDIA CUDA :: 12",
    "Topic :: Scientific/Engineering",
]
description = "A GPU-accelerated physics engine for robotics simulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "warp-lang>=1.11.0",
    "newton-actuators",
    "trimesh>=4.9.0",
    "scipy>=1.15.3",
    "gitpython>=3.1.45",
    "pycollada>=0.9.2",
    "mujoco>=3.5.0",
    "mujoco-warp>=3.5.0",
]

[project.optional-dependencies]

# Core simulation dependencies
sim = [
    # MuJoCo simulation
    "mujoco-warp>=3.5.0",         # MuJoCo, warp-accelerated
    "mujoco>=3.5.0",              # MuJoCo physics engine
]

# Asset import and mesh processing dependencies
importers = [
    # HTTP client for downloading meshes from URIs
    "requests>=2.25.0",

    # Mesh processing dependencies used by ModelBuilder.approximate_meshes
    "scipy",                       # for convex hull approximation
    "fast-simplification>=0.1.11", # for mesh simplification
    "alphashape>=1.3.1",           # for alphashape remeshing
    "coacd>=1.0.7",                # for convex decomposition
    "trimesh>=4.6.8",              # for mesh operations and V-HACD
    "meshio>=5.3.0",               # alternative mesh file loader (fallback)
    "pycollada>=0.9",              # transitive dependency of trimesh for COLLADA (.dae) mesh support
    "resolve-robotics-uri-py>=0.4.0",  # for resolving ROS/Gazebo-style URIs (package://, model://)

    # USD core libraries
    "usd-core>=25.5 ; platform_machine != 'aarch64'",
    "usd-exchange>=2.1.0a3 ; platform_machine == 'aarch64'",

    # Newton USD Schemas
    "newton-usd-schemas>=0.1.0rc3",
]

# Remeshing dependencies (for SurfaceReconstructor)
remesh = [
    "open3d>=0.18.0",              # for Poisson surface reconstruction
    "pyfqmr>=0.5.0",               # for fast mesh simplification
]

# Examples dependencies including visualization
examples = [
    "newton[sim]",          # include simulation dependencies
    "newton[importers]",    # include import dependencies
    "pyglet>=2.1.6",        # for OpenGL viewer
    "GitPython>=3.1.44",    # for downloading assets via newton.utils.download_asset()
    "imgui_bundle>=1.92.0", # for viewer GUI
    "pyyaml>=6.0.2",
    "cbor2>=5.7.0",         # for binary recording format (.bin files) - more efficient than JSON
]

# Extra dependency needed to run examples that inference an RL policy
torch-cu12 = [
    "newton[examples]", # include example dependencies
    "torch>=2.7.0",     # Cuda 12 PyTorch
]

# Development and Testing
# Contains all w/o torch, plus development debugging dependencies
dev = [
    "newton[examples]",      # includes examples, which pulls in sim and visualization
    "coverage[toml]>=7.8.0", # for code coverage testing
]

# Documentation dependencies
docs = [
    "newton[importers]",    # include import dependencies
    "myst-parser>=3.0.1",
    "sphinx>=7.4.7",
    "pydata_sphinx_theme>=0.16.1",
    "sphinxcontrib-mermaid>=1.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-design<=0.6.1",
    "sphinx-tabs>=3.3.0",
    "nbsphinx>=0.9.0",              # for Jupyter notebook integration
    "ipykernel>=6.0.0,<7.0.0",      # required by nbsphinx to execute notebooks (pinned <7 for rerun-sdk compatibility)
    "pypandoc>=1.11",               # required by nbsphinx for notebook conversion
    "pypandoc_binary>=1.11",        # bundled pandoc binary
    "viser>=1.0.16",                # required for ViewerViser in tutorial notebooks
]

# Dependencies for Jupyter notebook compatibility
notebook = [
    "newton[examples]",
    "rerun-sdk[notebook]>=0.27.1",
    "jupyterlab>=4.4.10",
]

[tool.hatch.version]
path = "newton/_version.py"

[tool.hatch.build.targets.wheel]
include = ["/newton", "/newton/licenses"]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com/"
explicit = true

[tool.uv.sources]
warp-lang = { index = "nvidia" }
torch = [{ index = "pytorch-cu128", extra = "torch-cu12" }]

[dependency-groups]
dev = ["asv>=0.6.4"]

[tool.ruff]
cache-dir = ".cache/ruff"
line-length = 120
indent-width = 4
exclude = ["*.js"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "I",   # isort
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "NPY", # NumPy-specific rules
    "PLC", # pylint convention
    "PLE", # pylint error
    "PLW", # pylint warning
    "UP",  # pyupgrade
    "RUF",  # ruff-specific rules
    "TID253",  # banned module-level imports
]
ignore = [
    "E501",   # Many lines are over 120 characters already (TODO: Fix)
    "E741",   # Warp often uses l as a variable name
    "UP018",  # Warp kernels often use float(val) and int(val)
    "RUF046", # Warp kernels often use int(val)
    "F811",   # Allow Warp kernel/function overloads
    "TC001",  # Allow quotations in type annotations, e.g. Literal["a"]
    "TC002",
    "TC003",
    "F821",
]

[tool.ruff.lint.per-file-ignores]
"asv/benchmarks/**/*" = ["RUF012", "E402"]
"newton/examples/**/*" = ["E402", "TID253"]
# TID253 (banned module-level imports) only applies to newton/_src and newton/tests
"docs/**/*" = ["TID253"]
"asv/**/*" = ["TID253"]

[tool.ruff.lint.flake8-tidy-imports]
banned-module-level-imports = [
    "alphashape",
    "coacd",
    "fast_simplification",
    "git",
    "imgui_bundle",
    "IPython",
    "matplotlib",
    "meshio",
    "mujoco",
    "mujoco_warp",
    "networkx",
    "open3d",
    "openmesh",
    "PIL",
    "point_cloud_utils",
    "pxr",
    "pyfqmr",
    "pyglet",
    "pytetwild",
    "requests",
    "scipy",
    "torch",
    "trimesh",
    "viser",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.coverage.run]
source = ["newton"]
disable_warnings = [
    "module-not-measured",
    "module-not-imported",
    "no-data-collected",
    "couldnt-parse",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@wp",
    "@warp",
    "if 0:",
    "if __name__ == .__main__.:",
]
omit = [
    "*/newton/examples/*",
    "*/newton/tests/*",
    "*/newton/tests/thirdparty/*",
]

[tool.basedpyright]
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportAny = false
reportExplicitAny = false
reportUnannotatedClassAttribute = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportOptionalOperand = false
reportUnusedCallResult = false

[tool.typos]
files.extend-exclude = ["/examples/assets", "*.urdf", "*.usd", "*.usda", "*.mjcf", "*.js"]

[tool.typos.default.extend-words]
arange = "arange"
ba = "ba"
HAA = "HAA"
lod = "lod"
# MuJoCo Menagerie brand/product names
Pincher = "Pincher"
Robotis = "Robotis"
robotis = "robotis"
Iit = "Iit"
IIT = "IIT"
iit = "iit"
PN = "PN"
pnd = "pnd"

[tool.typos.default.extend-identifiers]
PNGs = "PNGs"
Haa = "Haa"
